+++ src/singletons/Settings.hpp
@@ -555,6 +555,7 @@ public:
     BoolSetting informOnTabVisibilityToggle = {"/misc/askOnTabVisibilityToggle",
                                                true};
     BoolSetting lockNotebookLayout = {"/misc/lockNotebookLayout", false};
+    BoolSetting showPronouns = {"/misc/showPronouns", false};
 
     /// UI